backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
bcrypt==4.1.2
cryptography==41.0.7
stripe==7.8.0
sendgrid==6.11.0
playwright==1.40.0
APScheduler==3.10.4
pytz==2023.3
requests==2.31.0
gunicorn==21.2.0
